Add probe-domain entry point to return domain info
Add the `probe-domain` entry point, which is a command that returns key
record data for an target domain (SOA, NS, and A records, as well as the
IP-BGP resolution of resolved A records). Still pending is a WHOIS
lookup for the domain.

- Restructure package layout
- Add dependencies on [aslookup](https://github.com/dspruell/aslookup)
  and [whois-format](https://github.com/dspruell/whois-format)
- Add new probe-domain entry point
- Update tox config for new pathing
