+++ resolve_hosts/__init__.py
@@ -1,0 +1,1 @@
+"""Resolve list of DNS hostnames."""
