+++ tox.ini
@@ -26,7 +26,7 @@ deps =
     flake8
     flake8-docstrings
 commands =
-    flake8 --docstring-convention pep257 {posargs:resolve_hosts.py}
+    flake8 --docstring-convention pep257 {posargs:resolve_hosts}
 
 [testenv:type]
 description = run type checks
@@ -36,7 +36,7 @@ deps =
     types-tabulate
     types-ujson
 commands =
-    mypy {posargs:tests resolve_hosts.py}
+    mypy {posargs:tests resolve_hosts}
 
 [testenv:release]
 # specify Python 3 to use platform's default Python 3
